The libdime functions were updated to use the new prime interfaces.
Specifically to use the secp256k1 functions for key serialization
and deserialization. C++ support was also added to the magma 
project configuration.

+++ src/providers/cryptography/stacie.c
@@ -973,7 +973,7 @@ stringer_t * stacie_realm_decrypt(stringer_t *vector_key, stringer_t *tag_key, s
 	mm_move((uchr_t *)&pad, st_data_get(output) + 3, 1);
 	buffer_len = be32toh(buffer_len_be);
 
-	if (((buffer_len + pad + 4) % 16) != 0 || (buffer_len + pad + 4) != used_len) {
+	if (((buffer_len + pad + 4) % 16) != 0 || (buffer_len + pad + 4) !=  (size_t)used_len) {
 		log_pedantic("An invalid length or pad value was found inside the decrypted buffer.");
 		st_free(output);
 		return NULL;
